+++ includes/upgrades/upgrade-2.1.9.php
@@ -51,7 +51,7 @@ function wpuf_upgrade_2_1_9_subscription() {
         return;
     }
 
-    $post_type = (new \WeDevs\Wpuf\Subscription())->get_all_post_type();
+    $post_type = (new \WeDevs\Wpuf\Admin\Subscription())->get_all_post_type();
 
     foreach ( $results as $key => $result ) {
         $args = [
